(SIMP-287) Fix UTF-8/ASCII-8BIT encoding error
Before this patch, "simp config" would fail in EL7/Ruby 1.9+, due to a
difference in the encoding of salt (ASCII-8BIT, due to a workaround for
Ruby 1.8.7) and other Strings when concatenated inside Base64.encode64.

This patch forces the encoding for salt to UTF8 if the version of Ruby
is newer than 1.8, and bumps the gem and RPM release to 1.0.2.

SIMP-287 #close #comment Force encoding of salt to UTF-8 in ruby 2.0.0