Update patch set 3
Patch Set 3: Code-Review-1

Soooooo....what happens if we run 'simp config' later and want to re-enable ldap?

I see a remove function, but I'm not seeing an 'add' function.

We may want to just strip all of this out of Hiera by default and shove it down to the 'simp' module.

Patch-set: 3
Label: Code-Review=-1