(SIMP-497) Ensure path to fips_enabled fact
Before this commit, 'simp config' always detected (and suggested)
`use_fips=no` even if FIPS was enabled.  This patch ensures that even
before puppet is run, `simp config` will have access to the
`fips_enabled` fact by adding `common/lib/facter` to the Facter search
path.

NOTE: This fix is naive and direct.  It is intended to address only the
(common) scenario where `simp config` is run on a freshly-installed SIMP
OS.  It will not work if the ISO changes structure or on a non-SIMP OS.

SIMP-497 #close #comment Added direct path to common/lib/facter