SIMP-(6205) Ensure puppet.your.domain.yaml exists during `simp config`
The definition of @alt_file was modified to add '/' as the first
parameter in File.join(). This fixes the original issue of
puppet.your.domain.yaml not existing on a second run of `simp config`

This fix caused the potential for `simp config` to create multiple,
potentially valid, <host>.yaml files. Instead of deleting .yaml
files, they are backed up.

backup_host_yaml() was modified to accept a filename variable. The
module now detects all existing .yaml files and backs them up in the
same way that <host>.yaml was backed up before.

apply_summary was modified to report that extra <host>.yaml files exist
during the post `simp config` summary.

The number of .yaml files in environments/simp/data/hosts is currently
not reported in the summary because travis-ci fails when an extra line
is added.