(SIMP-6704) bootstrap cannot be rerun
Hardcoding the number of tagged runs caused bootstrap
to fail when it was rerun, because only 1 run is required
the second time around. To fix this problem, replaced the
hardcoded number of tagged runs with logic to ensure
bootstrap only executes these runs up to the point the
puppetserver is no longer running on the bootstrap port.

SIMP-6704 #close