Fix setting the hostname during 'simp config'
Always set the hostname to the one specified during `simp config` during
the apply phase.

This fixes an edge case where the user could decide on an alternate
hostname but choose not to configure the network interface. In this
case, the hostname would not be set properly and the resulting
`simp bootstrap` would fail to execute

Closes #182